requirements.txt
Pillow>=10.0
numpy>=1.24
scikit-image>=0.21
fonttools>=4.40
